+++ main/lib/idds/agents/carrier/poller.py
@@ -18,12 +18,11 @@ from idds.common.utils import setup_logging, truncate_string
 from idds.core import processings as core_processings
 from idds.agents.common.baseagent import BaseAgent
 from idds.agents.common.eventbus.event import (EventType,
-                                               NewProcessingEvent,
                                                UpdateProcessingEvent,
                                                SyncProcessingEvent,
                                                TerminatedProcessingEvent)
 
-from .utils import handle_new_processing, handle_update_processing, is_process_terminated
+from .utils import handle_update_processing, is_process_terminated
 
 setup_logging(__name__)
 
@@ -85,41 +84,6 @@ class Poller(BaseAgent):
         status = [ProcessingStatus.New, ProcessingStatus.Submitting, ProcessingStatus.Submitted,
                   ProcessingStatus.Running, ProcessingStatus.FinishedOnExec]
         core_processings.clean_next_poll_at(status)
-
-    def get_new_processings(self):
-        """
-        Get new processing
-        """
-        try:
-            if not self.is_ok_to_run_more_processings():
-                return []
-
-            self.show_queue_size()
-
-            processing_status = [ProcessingStatus.New]
-            processings = core_processings.get_processings_by_status(status=processing_status, locking=True,
-                                                                     not_lock=True,
-                                                                     new_poll=True, only_return_id=True,
-                                                                     bulk_size=self.retrieve_bulk_size)
-
-            # self.logger.debug("Main thread get %s [new] processings to process" % len(processings))
-            if processings:
-                self.logger.info("Main thread get [new] processings to process: %s" % str(processings))
-
-            for pr_id in processings:
-                self.logger.info("NewProcessingEvent(processing_id: %s)" % pr_id)
-                event = NewProcessingEvent(publisher_id=self.id, processing_id=pr_id)
-                self.event_bus.send(event)
-
-            return processings
-        except exceptions.DatabaseException as ex:
-            if 'ORA-00060' in str(ex):
-                self.logger.warn("(cx_Oracle.DatabaseError) ORA-00060: deadlock detected while waiting for resource")
-            else:
-                # raise ex
-                self.logger.error(ex)
-                self.logger.error(traceback.format_exc())
-        return []
 
     def get_running_processings(self):
         """
@@ -189,69 +153,6 @@ class Poller(BaseAgent):
                                                                      processing['transform_id'],
                                                                      processing['processing_id'])
 
-    def handle_new_processing(self, processing):
-        try:
-            log_prefix = self.get_log_prefix(processing)
-
-            # transform_id = processing['transform_id']
-            # transform = core_transforms.get_transform(transform_id=transform_id)
-            # work = transform['transform_metadata']['work']
-            status, processing, new_contents, msgs, errors = handle_new_processing(processing,
-                                                                                   self.agent_attributes,
-                                                                                   logger=self.logger,
-                                                                                   log_prefix=log_prefix)
-
-            if not status:
-                raise exceptions.ProcessSubmitFailed(errors)
-
-            parameters = {'status': ProcessingStatus.Submitting,
-                          'locking': ProcessingLocking.Idle,
-                          'processing_metadata': processing['processing_metadata']}
-            parameters = self.load_poll_period(processing, parameters)
-
-            proc = processing['processing_metadata']['processing']
-            if proc.submitted_at:
-                if not processing['submitted_at'] or processing['submitted_at'] < proc.submitted_at:
-                    parameters['submitted_at'] = proc.submitted_at
-
-            # if processing['processing_metadata'] and 'processing' in processing['processing_metadata']:
-            if proc.workload_id:
-                parameters['workload_id'] = proc.workload_id
-
-            update_processing = {'processing_id': processing['processing_id'],
-                                 'parameters': parameters}
-            ret = {'update_processing': update_processing,
-                   'update_contents': [],
-                   'new_contents': new_contents,
-                   'messages': msgs,
-                   }
-        except Exception as ex:
-            self.logger.error(ex)
-            self.logger.error(traceback.format_exc())
-
-            retries = processing['new_retries'] + 1
-            if not processing['max_new_retries'] or retries < processing['max_new_retries']:
-                pr_status = processing['status']
-            else:
-                pr_status = ProcessingStatus.Failed
-            # increase poll period
-            new_poll_period = int(processing['new_poll_period'].total_seconds() * self.poll_period_increase_rate)
-            if new_poll_period > self.max_new_poll_period:
-                new_poll_period = self.max_new_poll_period
-
-            error = {'submit_err': {'msg': truncate_string('%s' % (ex), length=200)}}
-            parameters = {'status': pr_status,
-                          'new_poll_period': new_poll_period,
-                          'errors': processing['errors'] if processing['errors'] else {},
-                          'new_retries': retries}
-            parameters['errors'].update(error)
-
-            update_processing = {'processing_id': processing['processing_id'],
-                                 'parameters': parameters}
-            ret = {'update_processing': update_processing,
-                   'update_contents': []}
-        return ret
-
     def update_processing(self, processing, processing_model):
         try:
             if processing:
@@ -313,31 +214,6 @@ class Poller(BaseAgent):
                 self.logger.error(ex)
                 self.logger.error(traceback.format_exc())
 
-    def process_new_processing(self, event):
-        self.number_workers += 1
-        try:
-            if event:
-                # pr_status = [ProcessingStatus.New]
-                self.logger.info("process_new_processing, event: %s" % str(event))
-                pr = self.get_processing(processing_id=event._processing_id, status=None, locking=True)
-                if not pr:
-                    self.logger.error("Cannot find processing for event: %s" % str(event))
-                else:
-                    log_pre = self.get_log_prefix(pr)
-                    self.logger.info(log_pre + "process_new_processing")
-                    ret = self.handle_new_processing(pr)
-                    # self.logger.info(log_pre + "process_new_processing result: %s" % str(ret))
-
-                    self.update_processing(ret, pr)
-
-                    # self.logger.info(log_pre + "SyncProcessingEvent(processing_id: %s)" % pr['processing_id'])
-                    # event = SyncProcessingEvent(publisher_id=self.id, processing_id=pr['processing_id'])
-                    # self.event_bus.send(event)
-        except Exception as ex:
-            self.logger.error(ex)
-            self.logger.error(traceback.format_exc())
-        self.number_workers -= 1
-
     def handle_update_processing(self, processing):
         try:
             log_prefix = self.get_log_prefix(processing)
@@ -346,11 +222,13 @@ class Poller(BaseAgent):
                                                                                                logger=self.logger,
                                                                                                log_prefix=log_prefix)
 
+            new_process_status = process_status
             if is_process_terminated(process_status):
-                process_status = ProcessingStatus.Terminating
+                new_process_status = ProcessingStatus.Terminating
 
             update_processing = {'processing_id': processing['processing_id'],
-                                 'parameters': {'status': process_status,
+                                 'parameters': {'status': new_process_status,
+                                                'substatus': process_status,
                                                 'locking': ProcessingLocking.Idle}}
 
             update_processing['parameters'] = self.load_poll_period(processing, update_processing['parameters'])
@@ -370,7 +248,7 @@ class Poller(BaseAgent):
                    'update_contents': update_contents,
                    'new_contents': new_contents,
                    'messages': ret_msgs,
-                   'processing_status': process_status}
+                   'processing_status': new_process_status}
 
         except exceptions.ProcessFormatNotSupported as ex:
             self.logger.error(ex)
@@ -460,10 +338,6 @@ class Poller(BaseAgent):
 
     def init_event_function_map(self):
         self.event_func_map = {
-            EventType.NewProcessing: {
-                'pre_check': self.is_ok_to_run_more_processings,
-                'exec_func': self.process_new_processing
-            },
             EventType.UpdateProcessing: {
                 'pre_check': self.is_ok_to_run_more_processings,
                 'exec_func': self.process_update_processing
@@ -483,9 +357,6 @@ class Poller(BaseAgent):
             self.add_default_tasks()
 
             self.init_event_function_map()
-
-            task = self.create_task(task_func=self.get_new_processings, task_output_queue=None, task_args=tuple(), task_kwargs={}, delay_time=60, priority=1)
-            self.add_task(task)
 
             task = self.create_task(task_func=self.get_running_processings, task_output_queue=None, task_args=tuple(), task_kwargs={}, delay_time=60, priority=1)
             self.add_task(task)
